Broke out Search component to use in multiple places and update live
