+++ lib/App.js
@@ -23,7 +23,7 @@ const currDay = wData.forecast.simpleforecast.forecastday[0];
 export default class App extends Component {
     constructor() {
         super();
-        this.state = {}
+        this.state = {};
     }
 
     render() {
