+++ lib/Header.js
@@ -1,24 +1,16 @@
 import React from 'react';
+import Search from './Search.js'
 import './styles/Header.css';
 
 const Header = (props) => {
-  function getLoc(e) {
-    e.preventDefault();
-    let input = document.getElementById("search").value;
-    document.getElementById("search").value = "";
-  }
   return (
     <header className='Header'>
       <div className="header-info">
         <h3 className='location'>{props.location}</h3>
         <h3 className='logo'>Weatherly</h3>
       </div>
-      <form>
-      <input id="search" type="text" placeholder="search location" />
-        <button id="search-btn" onClick={getLoc}>
-          submit
-        </button>
-      </form>
+      <Search />
+      
     </header>
     )
 }
