+++ lib/styles/Search.css
@@ -1,0 +1,38 @@
+input:focus,
+select:focus,
+textarea:focus,
+button:focus {
+    outline: none;
+}
+
+form {
+  margin: 25px 50px;
+  display: inline-block;
+}
+
+#search {
+  box-sizing: border-box;
+  height: 35px;
+  background: none;
+  border: none;
+  border-bottom: 2px solid #ffe47a;
+  padding-bottom: 5px;
+  color: #ffe47a;
+  font-size: 20px;
+}
+
+#search::placeholder {
+  color: #ffe47a;
+}
+
+#search-btn {
+  font-size: 20px;
+  color: #ffe47a;
+  box-sizing: border-box;
+  margin-left: 5px;
+  border: none;
+  padding-bottom: 5px;
+  border-bottom: 2px solid #ffe47a;
+  background: none;
+  height: 35px;
+}
